+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#4637](https://github.com/xmake-io/xmake/issues/4637): Add mix generator for xpack
+* [#5107](https://github.com/xmake-io/xmake/issues/5107): Add deb generator for xpack
 * [#5148](https://github.com/xmake-io/xmake/issues/5148): Add on_source in package
 
 ### Changes
@@ -1828,6 +1829,7 @@
 ### 新特性
 
 * [#4637](https://github.com/xmake-io/xmake/issues/4637): 为 xpack 添加 mix 支持
+* [#5107](https://github.com/xmake-io/xmake/issues/5107): 为 xpack 添加 deb 支持
 * [#5148](https://github.com/xmake-io/xmake/issues/5148): 为包添加 on_source 配置域
 
 ### 改进